# Improve the performance calling c++