This PR speeds up c++ calls.
It adds unsafe call to c++ calls. This saves about 100ns per call.
And it reduces the number of times to allocate the data area for receiving exceptions and shorten it by about 100ns.
In my use case, it changes the cost per call from 250ns to 50ns.