@bitonic 2022-10-26
I'll put a release out this evening.

On Wed, 26 Oct 2022 at 07:56, junji hashimoto ***@***.***>
wrote:

> Thank you, too.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/fpco/inline-c/pull/140#issuecomment-1291583358>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEHYORVLSXEDCNY6PRSL2LWFDIY5ANCNFSM6AAAAAARMTGOHQ>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>